+++ custom_components/hacs_hunter_douglas_blind/__init__.py
@@ -8,7 +8,7 @@ from homeassistant.core import HomeAssistant
 
 from .hub import PowerViewHub
 
-PLATFORMS = [Platform.SENSOR]
+PLATFORMS = [Platform.SENSOR, Platform.BUTTON]
 
 type PowerViewConfigEntry = ConfigEntry[PowerViewHub]